Changed Pools so they are stored as a raw pointer rather then a shared pointer.

+++ Engine/include/Common/CedarTime.h
@@ -5,7 +5,7 @@ namespace cedar
 {
 	namespace Time
 	{
-		inline float FixedDeltaTime = FIXED_DT;
+		inline float FixedDeltaTime = static_cast<float>(FIXED_DT);
 		inline float DeltaTime      = 0.0f;
 		inline float AlphaTime      = 0.0f;
 	};
